Adds script for customers to analyze MDT IO hotspotting on a FSxL DNE file system
This script will be provided to FSxL customers that are using DNE file system. The purpose of this script is to identify which files/directories on a given MDT host were most recently accessed by the customer workload. Based on the data outputted by the script, the customer can take action to perform a passive rebalancing in order to balance IO across multiple MDT hosts.

The script generates an output csv file that contains the number of files accessed by the workload in each directory and the total number of files in each directory. The customers can use this information to decide if they need to migrate a set of directories from one MDS host to another.